// src/types/rating-notification.types.ts
export enum RatingNotificationType {
  RATING_RECEIVED = 'RATING_RECEIVED',
  RATING_UPDATED = 'RATING_UPDATED',
  RATING_DISPUTED = 'RATING_DISPUTED',
  RATING_RESPONSE_ADDED = 'RATING_RESPONSE_ADDED',
  RATING_DISPUTE_RESOLVED = 'RATING_DISPUTE_RESOLVED',
  RATING_PUBLISHED = 'RATING_PUBLISHED',
  LOW_RATING_ALERT = 'LOW_RATING_ALERT',
  HIGH_RATING_ALERT = 'HIGH_RATING_ALERT',
  RATING_THRESHOLD_REACHED = 'RATING_THRESHOLD_REACHED'
} 